Fix display

+++ src/Display/Display.cpp
@@ -1,17 +1,17 @@
 #include "Display.h"
 
-void Display::set(int x, int y) {
+void Display::set(uint8_t x, uint8_t y) {
     pixels.at(x).at(y) = true;
 }
 
-void Display::unset(int x, int y) {
+void Display::unset(uint8_t x, uint8_t y) {
     pixels.at(x).at(y) = false;
 }
 
-void Display::flip(int x, int y) {
+void Display::flip(uint8_t x, uint8_t y) {
     pixels.at(x).at(y) = !pixels.at(x).at(y);
 }
 
-bool Display::getPixel(int x, int y) {
+bool Display::getPixel(uint8_t x, uint8_t y) {
     return pixels.at(x).at(y);
 }
